+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/util/tasks.kt
@@ -1,13 +1,12 @@
 package io.sentry.android.gradle.util
 
-import io.sentry.android.gradle.SentryTasksProvider.capitalized
 import io.sentry.android.gradle.SentryTasksProvider.getAssembleTaskProvider
 import io.sentry.android.gradle.SentryTasksProvider.getBundleTask
 import io.sentry.android.gradle.SentryTasksProvider.getInstallTaskProvider
+import io.sentry.android.gradle.SentryTasksProvider.getMinifyTask
 import io.sentry.android.gradle.SentryTasksProvider.getPackageBundleTask
 import io.sentry.android.gradle.SentryTasksProvider.getPackageProvider
 import io.sentry.android.gradle.SentryTasksProvider.getPreBundleTask
-import io.sentry.android.gradle.SentryTasksProvider.getTransformerTask
 import io.sentry.android.gradle.util.SentryPluginUtils.withLogging
 import io.sentry.gradle.common.SentryVariant
 import org.gradle.api.Project
@@ -22,28 +21,15 @@ fun TaskProvider<out Task>.hookWithMinifyTasks(
     // we need to wait for project evaluation to have all tasks available, otherwise the new
     // AndroidComponentsExtension is configured too early to look up for the tasks
     project.afterEvaluate {
-        val transformerTaskProvider = withLogging(project.logger, "transformerTask") {
-            getTransformerTask(
-                project,
-                variantName,
-                dexguardEnabled
-            )
-        }
+        val minifyTask = getMinifyTask(
+            project,
+            variantName,
+            dexguardEnabled
+        )
 
-        if (dexguardEnabled &&
-            GroovyCompat.isDexguardEnabledForVariant(project, variantName)
-        ) {
-            project.tasks.named(
-                "dexguardApk${variantName.capitalized}"
-            ).configure { it.finalizedBy(this) }
-            project.tasks.named(
-                "dexguardAab${variantName.capitalized}"
-            ).configure { it.finalizedBy(this) }
-        } else {
-            // we just hack ourselves into the Proguard/R8 task's doLast.
-            transformerTaskProvider?.configure {
-                it.finalizedBy(this)
-            }
+        // we just hack ourselves into the Proguard/R8/DexGuard task's doLast.
+        minifyTask?.configure {
+            it.finalizedBy(this)
         }
     }
 }